// Codigo/front-end/src/app/interfaces/Parent.ts
export interface Parent {
  fullName: string;
  cpf: string;
  rg: string;
  email: string;
  occupation: string;
  company: string;
  phoneNumber: string;
  streetAddress: string;
  addressNumber: string;
  neighborhood: string;
  addressComplement: string,
  cityOfResidence: string;
  zipCode: string;
  homeState: string;
  birthDate: string;
  relationship: string;
}